First version
This is the first version of a currency converter web app. It utilizes Exchangerate's API, enabling us to access a wide range of currencies from around the world. This version incorporates a simple design created using CSS and HTML. It does not rely on any frameworks or employ special effects... yet.